final 11.0

+++ check_wall.c
@@ -18,7 +18,7 @@ int		check_up(char **map, int x, int y)
 		return (-1);
 	while (y-- > 0)
 	{
-		if (x > ft_strlen(map[y]) - 1)
+		if (x > (int)ft_strlen(map[y]) - 1)
 			return (-1);
 		if (map[y][x] == ' ')
 			return (-1);
@@ -34,7 +34,7 @@ int		check_bottom(char **map, int x, int y, int length_map_y)
 		return (-1);
 	while (y++ < length_map_y - 1)
 	{
-		if (x > ft_strlen(map[y]) - 1)
+		if (x > (int)ft_strlen(map[y]) - 1)
 			return (-1);
 		if (map[y][x] == ' ')
 			return (-1);
@@ -60,7 +60,7 @@ int		check_left(char **map, int x, int y)
 
 int		check_right(char **map, int x, int y)
 {
-	if (x == ft_strlen(map[y]) - 1)
+	if (x == (int)ft_strlen(map[y]) - 1)
 		return (-1);
 	while (x-- > 0)
 	{
